+++ packages/core/__test__/injector/injector.spec.ts
@@ -1,6 +1,10 @@
 import { expect } from 'chai';
 import { Injector } from '../../injector';
-import { Module1 } from './mocks/classModuleToInject';
+import {
+  Module1,
+  ModuleFactory,
+  ModuleProviderError,
+} from './mocks/classModuleToInject';
 import {
   Controller1,
   Controller2,
@@ -9,9 +13,19 @@ import {
 } from './mocks/classDependenciesToInject';
 
 describe('Core - Injector - injector', () => {
-  const listProvider = new Map();
-  const injector = new Injector(Module1, listProvider);
-
+  let listProvider;
+  let listProviderClass;
+  let injector: Injector;
+  beforeEach(() => {
+    listProvider = new Map();
+    listProviderClass = new Map();
+    injector = null;
+    injector = new Injector(Module1, listProvider, listProviderClass);
+    injector.scanProviders();
+  });
+  after(() => {
+    injector = null;
+  });
   it('Should respond an object without paramters to inject', () => {
     const getInjectData = injector.getInjectData(Controller1);
     expect(getInjectData).to.is.empty;
@@ -42,29 +56,25 @@ describe('Core - Injector - injector', () => {
     const allProvider = injector.getAllProvider();
     expect(allProvider.has('TEXT_INJECT')).to.be.equal(true);
   });
-  it('Should be get a function as a provider', () => {
-    function getTrue() {
-      return true;
-    }
-    const getProvider = injector.getInjectProvider({
-      key: 'getTrueFunction',
-      value: getTrue,
-    });
-    expect(getProvider).to.have.property('resolve');
-  });
-  it('Should be set a object as a provider', () => {
-    const getProvider = injector.getInjectProvider({
-      key: 'getTrue',
-      value: 'true',
-    });
-    expect(getProvider).to.have.property('resolve');
-  });
 
-  it('should return the function that returns the given injection data', () => {
-    const injectData = {
-      Some: 'some some',
-    };
-    const resultFunction = injector.funtionInjectData(injectData);
-    expect(resultFunction()).to.be.equal(injectData);
+  it('Should respond an number type asFunction without injector', () => {
+    const injectorWithFactory = new Injector(
+      ModuleFactory,
+      listProvider,
+      listProviderClass,
+    );
+    injectorWithFactory.scanProviders();
+    const allProvider = injectorWithFactory.getAllProvider();
+    expect(allProvider.has('NUMBER_FACTORY')).to.be.equal(true);
+  });
+  it('Should respond only one supplier because the others are not valid.', () => {
+    const injectorWithFactory = new Injector(
+      ModuleProviderError,
+      listProvider,
+      listProviderClass,
+    );
+    injectorWithFactory.scanProviders();
+    const allProviderClass = injectorWithFactory['listProvidersClass'];
+    expect(allProviderClass.size).to.be.equal(1);
   });
 });