+++ packages/core/__test__/injector/mocks/classModuleToInject.ts
@@ -1,11 +1,16 @@
 import { Module } from '@zanobijs/common';
 import {
+  ABSServiceRepository,
   Controller1,
   Controller2,
   Controller3,
   Controller4,
   Controller5,
+  RepositoryImplements,
   Service1,
+  ServiceUseCase,
+  ServiceUseFactory,
+  ServiceWithoutInjectable,
 } from './classDependenciesToInject';
 
 // ==================================================
@@ -25,6 +30,20 @@ import {
   exports: [],
 })
 export class Module1 {}
+
+@Module({
+  imports: [],
+  controllers: [],
+  services: [
+    ServiceUseCase,
+    {
+      provider: ABSServiceRepository,
+      useClass: RepositoryImplements,
+    },
+  ],
+  exports: [],
+})
+export class ModuleRepository {}
 
 // ==================================================
 // ===========Module to Test Module =================
@@ -67,3 +86,69 @@ export class Module3 {}
   exports: [],
 })
 export class Module4 {}
+
+@Module({
+  imports: [],
+  controllers: [Controller5],
+  services: [
+    ServiceUseFactory,
+    {
+      provider: 'NUMBER_FACTORY',
+      useFactory: () => {
+        return 45;
+      },
+    },
+  ],
+  exports: [],
+})
+export class ModuleFactory {}
+
+@Module({
+  imports: [],
+  controllers: [Controller5],
+  services: [
+    ServiceUseFactory,
+    {
+      provider: Service1,
+      useFactory: () => {
+        return 45;
+      },
+    },
+    {
+      provider: Service1,
+      useValue: 'SOME TEXT',
+    },
+    {
+      provider: 'MyService1',
+      useClass: Service1,
+    },
+  ],
+  exports: [],
+})
+export class ModuleProviderError {}
+
+@Module({
+  imports: [],
+  controllers: [Controller5],
+  services: [
+    {
+      provider: 'MyService1',
+      useClass: ServiceWithoutInjectable,
+    },
+  ],
+  exports: [],
+})
+export class ModuleProviderWithoutInjectable {}
+
+@Module({
+  imports: [],
+  controllers: [],
+  services: [
+    {
+      provider: 'MyService1',
+      useClass: ModuleFactory,
+    },
+  ],
+  exports: [],
+})
+export class ModuleProviderHaveModule {}